Update to include padding between labels
I have not updated the length of the text in the various labels. I have updated the constraints to use simpler logic and added padding around and between the labels.

In situations where the length of labels dictates that one or both of them must be compressed, it appears that autolayout will always favour compressing the rightmost label as much as possible before compressing the leftmost label at all.